Use SEMI_AUTOMATIC mode

+++ firmware/go_photon.sh
@@ -2,11 +2,14 @@
 echo ------------------------------------------------------------------- build
 particle compile p . --saveTo firmware.bin
 
-echo ------------------------------------------------------------------- check usb
-sudo dfu-util -l
+if [ "$?" = "0" ]; then
+  echo ------------------------------------------------------------------- check usb
+  sudo dfu-util -l
 
-echo ------------------------------------------------------------------- flash
-sudo particle flash --usb firmware.bin
-rm firmware.bin
-
+  echo ------------------------------------------------------------------- flash
+  sudo particle flash --usb firmware.bin
+  rm firmware.bin
+else
+  echo build failed
+fi
